requirements.txt
requests
pandas
dash
plotly
fuzzywuzzy
python-Levenshtein
python-dotenv
